% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_logisitcal_learning.R
\name{generate_predictions}
\alias{generate_predictions}
\title{Generate Predictions Using a Logistic Model}
\usage{
generate_predictions(logistic_model, new_data, type = "response")
}
\arguments{
\item{logistic_model}{A logistic regression model object, typically created
and trained using \code{\link[stats]{glm}} with a binomial family. This
model is used to generate predictions on the new data.}

\item{new_data}{A data frame or matrix containing the new observations for
which predictions are to be made. The structure of this data should match
the model's training data, excluding the dependent variable.}

\item{type}{Character string specifying the type of prediction required.
Default is "response", which returns predicted probabilities. Other options
depend on the \code{\link[stats]{predict.glm}} method, such as "link" for
the linear predictors.}
}
\value{
A numeric vector of predicted probabilities for the positive class
  (assuming a binary classification problem) if \code{type = "response"} is
  used. The length of the vector matches the number of rows in
  \code{new_data}. For other values of \code{type}, the return value follows
  the specification of \code{\link[stats]{predict.glm}}.
}
\description{
This function generates predictions from a logistic regression model for a
new dataset. It assumes the model is already fitted using logistic regression
and the new data is prepared in a similar format as the training data. The
function is designed to work with binary classification problems.
}
