% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_logisitcal_learning.R
\name{get_cls_infos}
\alias{get_cls_infos}
\title{Calculate Classification Metrics and Optimal Cutoff}
\usage{
get_cls_infos(true_ones, preds, thrsh)
}
\arguments{
\item{true_ones}{Numeric vector of true class labels (0 or 1).}

\item{preds}{Numeric vector of predicted probabilities for the positive class.}

\item{thrsh}{Numeric value specifying the threshold for classification.}
}
\value{
A list containing the optimal cutoff, concordance, misclassification
error, sensitivity, and one minus specificity.
}
\description{
Computes various classification metrics including the optimal cutoff for
classification, misclassification error, concordance, sensitivity, and
specificity based on true outcomes and predicted probabilities.
}
