% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{get_data_for_summary_plot}
\alias{get_data_for_summary_plot}
\title{Binds (rbind()'s) data frames for different years}
\usage{
get_data_for_summary_plot(...)
}
\arguments{
\item{...}{data sets as produced by [get_data_summary()] through its second
list output element; but can be, more generally, any number of data sets of
the same kind (i.e. the same col-names/-types)}
}
\value{
\code{rbind}-ed data frames
}
\description{
Binds (rbind()'s) data frames for different years
}
