% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_summaries.R
\name{get_data_summary}
\alias{get_data_summary}
\title{Computes various data summaries from raw data}
\usage{
get_data_summary(data_segm, year, type = "all")
}
\arguments{
\item{data_segm}{a \code{data.frame} or \code{tibble} of segmented data (this
is typically the raw data set being run through [segmentation_analysis()]}

\item{year}{the year as a numeric variable}

\item{type}{the type of data set to return; if 'all' then a list of all three
as described in "Value" are returned, otherwise for "report", "divisions",
or "final", only the corresponding subtype is returned}
}
\value{
a list of three elements: \itemize{
    \item \code{df_final_data} the final data set
    \item \code{df_final_report}
    \item \code{df_final_divisions}
  }
  see the details section, for more information on each data set.
}
\description{
Data summaries include
  \itemize{
    \item \code{df_final_data} the final data set
    \item \code{df_final_report}
    \item \code{df_final_divisions}
  }
}
