% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{get_data_summary_radar}
\alias{get_data_summary_radar}
\title{Generates data set for [radar_plot()]}
\usage{
get_data_summary_radar(data_set)
}
\arguments{
\item{data_set}{data set, joined with all years, as produced by
[get_data_for_summary_plot()]}
}
\value{
a data frame, see function body for details
}
\description{
Generates data set for [radar_plot()]
}
