% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{get_pie_figures}
\alias{get_pie_figures}
\title{Main function to produce the pie figure}
\usage{
get_pie_figures(df, ar)
}
\arguments{
\item{df}{a data.frame as returned as the first element from the output list
from [get_data_summary()]}

\item{ar}{the year as an integer (2021, 2022, or 2023)}
}
\value{
a "grid arranged" ggplot
}
\description{
Main function to produce the pie figure
}
