% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisitcal_learning.R
\name{get_true_ones}
\alias{get_true_ones}
\title{Extract True Class Labels from Prediction Data Set}
\usage{
get_true_ones(data_set_predictions)
}
\arguments{
\item{data_set_predictions}{A data frame containing the predictions, with the
first column being the dependent variable.}
}
\value{
A numeric vector containing the true class labels as 0 or 1.
}
\description{
This function extracts the true class labels as binary outcomes (0 or 1)
from a given prediction data set. It is assumed that the first column of the
data set contains the dependent variable with values coded accordingly.
}
