% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisitcal_learning.R
\name{logistic_learn}
\alias{logistic_learn}
\title{Function wrapper for logit regressions adjusted to the SSV setting.}
\usage{
logistic_learn(data_set, model)
}
\arguments{
\item{data_set}{the data set; will be passed to get_data_for_prediction
internally}

\item{model}{a model specification of the form ....}
}
\value{
output as generated via \link[stats]{glm} but flavoured with some
  nicer printing structure and addtional information
}
\description{
The wrapper is araound \link[stats]{glm} and a suitable expression to be
passed to \code{as.formula} i.e. of the form "y~x1+x2+...+xn"
}
