% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03D_plot_divisions.R
\name{plot_division}
\alias{plot_division}
\title{Generate radar-type plot}
\usage{
plot_division(
  data_set,
  year,
  title,
  y_title = "Digital kompetanse i divisjonene",
  col_scm = c(light_blue = "#189BC4", dark_blue = "#3F505A", orange = "#F59331",
    light_green = "#54B64E", green = "#58B02C", dark_green = "#356A1A", black1 =
    "#1F282D", black2 = "#263036")
)
}
\arguments{
\item{data_set}{a \code{data.frame} as returned as the first element from the
output list from [get_data_summary()] (if set to type = "all") or the
\code{data.frame} directly if type = 'final'}

\item{year}{the year as an integer (2021, 2022, or 2023)}

\item{title}{title of plot}

\item{y_title}{sub-title of y-axis names (which consist of year and
`y_title`)}

\item{col_scm}{the col-scheme; default is the task color scheme as hard coded}
}
\value{
a ggplot2-object of the radar plot type
}
\description{
Generate radar-type plot
}
