% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{plot_summary}
\alias{plot_summary}
\title{Produces a summary plot}
\usage{
plot_summary(
  data_all,
  col_scm = c(light_blue = "#189BC4", dark_blue = "#3F505A", orange = "#F59331",
    light_green = "#54B64E", green = "#58B02C", dark_green = "#356A1A", black1 =
    "#1F282D", black2 = "#263036")
)
}
\arguments{
\item{data_all}{data set, joined with all years, as produced by
[get_data_for_summary_plot()]}

\item{col_scm}{the col-scheme; default is the task color scheme as hard coded}
}
\value{
a ggplot2-object, the summary plot
}
\description{
Horizontal bars showing, per year, the different levels of digital expertise
as percentages of the total.
}
