% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_segmentation_HELPERS.R
\name{recode_q16}
\alias{recode_q16}
\title{Data segmentation based on Q16}
\usage{
recode_q16(
  data_set,
  from_vals = c(3, 4),
  cut_grun = 4,
  cut_vide = 2,
  cut_avan = 2,
  verbose = FALSE
)
}
\arguments{
\item{data_set}{data set (either raw or already added some segmentatin vars)}

\item{from_vals}{numeric vector giving which values to recode from}

\item{cut_grun}{numeric cutoff value for summation score of "grunlegende"; if
summation score larger than `cut_grun`, then employee has grunlegende
skills}

\item{cut_vide}{numeric cutoff value for summation score of "videregaende";
if summation score larger than `cut_vide`, then employee has videregaende
skills}

\item{cut_avan}{numeric cutoff value for summation score of "avansert"; if
summation score larger than `cut_avan`, then employee has avansert skills}

\item{verbose}{used to print additional information, now deprecated}
}
\value{
a data set with additional variables, most importantly
  "kat_kommunikasjon" which is a factor:
  \itemize{
  \item "Uerfaren": 0
  \item "Grunnleggende": 1
  \item "Videregående": 2
  \item "Avansert": 3
  }
}
\description{
Generate data set with segmentation variables for "Communication and
Interaction".
}
