% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_segmentation_HELPERS.R
\name{recode_q19}
\alias{recode_q19}
\title{Data segmentation based on Q19}
\usage{
recode_q19(
  data_set,
  from_vals = c(3, 4),
  sum_score_val_grun = 2,
  sum_score_val_vide = 5,
  sum_score_val_avan = 3
)
}
\arguments{
\item{data_set}{data set (either raw or already added some segmentatin vars)}

\item{from_vals}{numeric vector giving which values to recode from}

\item{sum_score_val_grun}{numeric score value for summation score of
"grunnleggende"; if summation score larger/larger than/equal
`sum_score_val_grun`, then employee has grunnleggende skills}

\item{sum_score_val_vide}{numeric score value for summation score of
"videregaende"; if summation score larger/larger than/equal
`sum_score_val_vide`, then employee has videregaende skills}

\item{sum_score_val_avan}{numeric score value for summation score of
"avansert"; if summation score larger/larger than/equal
`sum_score_val_avan`, then employee has avansert skills}
}
\value{
a data set with additional variables, most importantly
  "kat_utstyr1" which is a factor:
  \itemize{
  \item "Uerfaren": 0
  \item "Grunnleggende": 1
  \item "Videregående": 2
  \item "Avansert": 3
  }
}
\description{
Generate data set with segmentation variables for "Proficiency in using
technology".
}
