% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_segmentation_HELPERS.R
\name{recode_qXX_cats}
\alias{recode_qXX_cats}
\title{Re-codes a set of value of a have-factor variable (with levels)}
\usage{
recode_qXX_cats(data_set, q_names, list_recodes, new_names, SETTINGS_FACT)
}
\arguments{
\item{data_set}{the data set containing the variable to re_code}

\item{q_names}{character giving the name of the variable to re-code}

\item{list_recodes}{a list of elements of \code{dynamic_dots}-type that will
be passed down to [dplyr::recode_factor()] and gives the from/to value
re-coding as required by this function}

\item{new_names}{character vector with same lengths as \code{q_names} and
\code{list_recodes} giving the new variable names of the resulting
indicators}

\item{SETTINGS_FACT}{a (named!) list with the following logical values:
\itemize{
  \item{\code{ADJUST_LABELS:}}{logical; if \code{TRUE} then
  code{class(var_type) -> 'haven_labelled', 'vctrs_vctr', 'double'} labels
  are adjusted i.e the \code{from}'th label is removed}
  \item{\code{AS_FACTOR:}}{logical; if \code{TRUE}, then conversion to a
  factor is performed}
  \item{\code{ORDERED:}}{logical; if \code{TRUE} (and if \code{AS_FACTOR =
  TRUE}), then the factor will be ordered}
  \item{LABELS}{logical; if \code{TRUE} (and if \code{AS_FACTOR =
  TRUE}), then factor labels are set}
}}
}
\value{
the same data set but with the re-coded variable(s)
}
\description{
The variable types permitted are of \code{class(var_type) ->
'haven_labelled', 'vctrs_vctr', 'double'} and \code{typeof(var_type) ->
double} and can be e.g. \code{Q16r1-Q16r11, Q17r1-Q17r10, Q14r1-Q14r8,
Q19r1-Q19r11, Q36, Q37, Q38} etc. The levels e.g. 5 (which corresponds to
"Vet ikke" i.e. "Do not know") and 1 (which corresponds to "Ikke i det hele
tatt" or similar, which means some form of "no"/'not at all") are often made
zero, whereas higher values like 2,3 and 4 are then mapped to 1. So this
functions creates a categorical or indicator variable
}
