% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_segmentation_HELPERS.R
\name{recode_qXX_rVals}
\alias{recode_qXX_rVals}
\title{Re-codes value of a variable (withlevels) into another value}
\usage{
recode_qXX_rVals(
  data_set,
  q_names,
  from = 5,
  to = 1,
  na_replacement = NULL,
  SETTINGS_FACT = list(ADJUST_LABELS = FALSE, AS_FACTOR = FALSE, ORDERED = FALSE,
    ADD_LABELS = FALSE)
)
}
\arguments{
\item{data_set}{the data set containing the variable to re_code}

\item{q_names}{character giving the name of the variable to re-code}

\item{from}{the value to be changed}

\item{to}{the new value to be assigned (instead of the one passed via
\code{from})}

\item{na_replacement}{value to replace \code{NA} values with; defaults to
\code{NULL} but must be set to a numeric value is \code{NA}'s are present,
otherwise there will be an error}

\item{SETTINGS_FACT}{a (named!) list with the following logical values:
\itemize{
  \item{\code{ADJUST_LABELS:}}{logical; if \code{TRUE} then
  \code{class(var_type) -> 'haven_labelled', 'vctrs_vctr', 'double'} labels
  are adjusted to the length of actual levels of the variable after
  re-coding took place}
  \item{\code{AS_FACTOR:}}{logical; if \code{TRUE}, then conversion to a
  factor is performed}
  \item{\code{ORDERED:}}{logical; if \code{TRUE} (and if \code{AS_FACTOR =
  TRUE}), then the factor will be ordered}
  \item{ADD_LABELS}{logical; if \code{TRUE} (and if \code{AS_FACTOR =
  TRUE}), then factor labels are set}
}}
}
\value{
the same data set but with the re-coded variable(s)
}
\description{
The variable types are of
\code{class(var_type) -> 'haven_labelled', 'vctrs_vctr', 'double'} and
\code{typeof(var_type) -> double} and can be e.g. \code{Q16r1-Q16r11,
Q17r1-Q17r10, Q14r1-Q14r8, Q19r1-Q19r11, Q36, Q37, Q38} etc. The level e.g.
5 (which corresponds to "Vet ikke" i.e. "Do not know") is often re-coded to
1 (which corresponds to "Ikke i det hele tatt" or similar, which corresponds
to some form of "no"/'not at all").
}
