% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskAnalyticsTB.R
\docType{package}
\name{TaskAnalyticsTB}
\alias{TaskAnalyticsTB}
\title{Summary of what the package does.}
\description{
Summary of what the package does.
}
\section{Summary of what the package does.}{
NA
}

