% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03C_plot_radar.R
\name{get_data_summary_radar}
\alias{get_data_summary_radar}
\title{Generates summary data set for radar-type plot}
\usage{
get_data_summary_radar(data_set)
}
\arguments{
\item{data_set}{a data set}
}
\value{
a \code{data.frame} for [plot_radar()]
}
\description{
Generates summary data set for radar-type plot
}
