% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/42_logisitcal_learning.R
\name{logistic_learn}
\alias{logistic_learn}
\title{Function wrapper for logit regressions adjusted to the SSV setting.}
\usage{
logistic_learn(data_set = NULL, model = NULL, type = "default")
}
\arguments{
\item{data_set}{the data set; will be passed to get_data_for_prediction
internally}

\item{model}{a model specification as a list containing named elements of
  the form
\itemize{
  \item 'dependent': the name of the dependent variable as a character string
  \item 'regressors': a character vector of regressor names
  \item 'experience': a two element vector giving the dependent
  variable levels (values coded to zero and one) that are being compared
 }}

\item{type}{Character string specifying the type of logistic regression
execution. "default" for a standard logistic regression model fitting and
summary, "shinyDB" for a Shiny app-specific version that includes
additional summaries like pseudo R-squared and odds ratios. The function
will stop with an error if an unrecognized type is provided.}
}
\value{
output as generated via \link[stats]{glm} but flavored with some
  nicer printing structure and additional information; in the
  \code{type = "default"} version a list of three elements, in the shiny
  version a list of two
}
\description{
The wrapper is araound \link[stats]{glm} and a suitable expression to be
passed to \code{as.formula} i.e. of the form "y~x1+x2+...+xn"
}
