% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plots.R
\name{plot_pie_figures}
\alias{plot_pie_figures}
\title{Main function to produce the pie figure}
\usage{
plot_pie_figures(df, year)
}
\arguments{
\item{df}{a data.frame as returned as the first element from the output list
from [get_data_summary()]}

\item{year}{the year as an integer (2021, 2022, or 2023)}
}
\value{
a "grid arranged" ggplot2 object
}
\description{
Main function to produce the pie figure
}
