% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_segmentation_HELPERS.R
\name{recode_q16}
\alias{recode_q16}
\title{Data segmentation based on Q16}
\usage{
recode_q16(
  data_set,
  from_vals = c(3, 4),
  sum_score_val_grun = 4,
  sum_score_val_vide = 2,
  sum_score_val_avan = 2,
  type_val_grun = "larger-equal",
  type_val_vide = "larger-equal",
  type_val_avan = "equal"
)
}
\arguments{
\item{data_set}{data set (either raw or already added some segmentatin vars)}

\item{from_vals}{numeric vector giving which values to recode from}

\item{sum_score_val_grun}{numeric score value for summation score of
"grunnleggende"; if summation score larger/larger than/equal
`sum_score_val_grun`, then employee has grunnleggende skills}

\item{sum_score_val_vide}{numeric score value for summation score of
"videregaende"; if summation score larger/larger than/equal
`sum_score_val_vide`, then employee has videregaende skills}

\item{sum_score_val_avan}{numeric score value for summation score of
"avansert"; if summation score larger/larger than/equal
`sum_score_val_avan`, then employee has avansert skills}

\item{type_val_grun}{character string: either "larger-equal" or "equal" to
indicate how to apply the sum_score threshold (allow larger/equal to get 1
or only equality to that sum_score to get a 1 indicator, zero else)}

\item{type_val_vide}{character string: either "larger-equal" or "equal" to
indicate how to apply the sum_score threshold (allow larger/equal to get 1
or only equality to that sum_score to get a 1 indicator, zero else)}

\item{type_val_avan}{character string: either "larger-equal" or "equal" to
indicate how to apply the sum_score threshold (allow larger/equal to get 1
or only equality to that sum_score to get a 1 indicator, zero else)}
}
\value{
a data set with additional variables, most importantly
  "kat_kommunikasjon" which is a factor:
  \itemize{
  \item "Uerfaren": 0
  \item "Grunnleggende": 1
  \item "Videregående": 2
  \item "Avansert": 3
  }
}
\description{
Generate data set with segmentation variables for "Communication and
Interaction".
}
