% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_segmentation.R
\name{segmentation_analysis}
\alias{segmentation_analysis}
\title{Perform a segmentation analysis on the raw data}
\usage{
segmentation_analysis(data_SVV)
}
\arguments{
\item{data_SVV}{raw data (as a \code{data.frame})}
}
\value{
data set after segmentation analysis
}
\description{
Perform a segmentation analysis on the raw data
}
